Feature: move config to npm
This turns the config into a npm dependency, its:

- cleans up the gulpfile.js
- reduces copy-paste hell
- allows config versions to be tied to vf-core versions

To do this I've also added the `./tools` directory to be a lerna directory.

This means we can make more parts of vf-core npm components, like the component generator.
