+++ gulpfile.js
@@ -1,48 +1,11 @@
 'use strict';
 
-const gulp = require('gulp');
-
-// -----------------------------------------------------------------------------
-// Configuration
-// -----------------------------------------------------------------------------
-
-// Pull in some optional configuration from the package.json file, a la:
-// "vfConfig": {
-//   "vfName": "My Component Library",
-//   "vfNameSpace": "myco-",
-//   "vfComponentPath": "./src/components",
-//   "vfComponentDirectories": [
-//      "vf-core-components",
-//      "../node_modules/your-optional-collection-of-dependencies"
-//     NOTE: Don't forget to symlink: `cd components` `ln -s ../node_modules/your-optional-collection-of-dependencies`
-//    ],
-//   "vfBuildDestination": "./build",
-//   "vfThemePath": "@frctl/mandelbrot"
-// },
-// all settings are optional
-// todo: this could/should become a JS module
 const fs = require('fs');
 const path = require('path');
-const config = JSON.parse(fs.readFileSync('./package.json'));
-// Load the vf-core package.json config
-let vfCoreConfig;
-if (config.name === '@visual-framework/vf-core') {
-  // if being run from within the vf-core project, use the local package.json
-  vfCoreConfig = JSON.parse(fs.readFileSync('./package.json'));
-} else {
-  // load vfCoreConfig from node_modules
-  vfCoreConfig = JSON.parse(fs.readFileSync(require.resolve('@visual-framework/vf-core/package.json')));
-}
-config.vfConfig = config.vfConfig || [];
-global.vfName = config.vfConfig.vfName || "Visual Framework 2.0";
-global.vfNamespace = config.vfConfig.vfNamespace || "vf-";
-global.vfComponentPath = config.vfConfig.vfComponentPath || path.resolve('.', __dirname + '/components');
-global.vfBuildDestination = config.vfConfig.vfBuildDestination || __dirname + '/temp/build-files';
-global.vfThemePath = config.vfConfig.vfThemePath || './tools/vf-frctl-theme';
-global.vfVersion = vfCoreConfig.version || 'not-specified';
-const componentPath = path.resolve('.', global.vfComponentPath).replace(/\\/g, '/');
-const componentDirectories = config.vfConfig.vfComponentDirectories || ['vf-core-components'];
-const buildDestionation = path.resolve('.', global.vfBuildDestination).replace(/\\/g, '/');
+const gulp = require('gulp');
+
+// Pull in some optional configuration from the package.json file, a la:
+const {componentPath, componentDirectories, buildDestionation} = require('./tools/vf-config');
 
 // Gulp tasks live in their own files, for the sake of clarity.
 // These are done as JS Modules as it makes passing paramaters simpler and avoids
