+++ lerna.json
@@ -1,7 +1,7 @@
 {
-  "lerna": "0.0.1",
   "packages": [
-    "components/**/**"
+    "components/**/**",
+    "tools/*"
   ],
   "version": "independent"
 }
