+++ tools/vf-config/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "vf-config",
+  "version": "0.0.1",
+  "description": "Pull in some optional configuration from a project's package.json for use when building Visual Framework components and templates.",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/visual-framework/vf-core.git"
+  },
+  "keywords": [
+    "Visual",
+    "Framework",
+    "VF"
+  ],
+  "author": "Ken Hawkins <ken.hawkins@embl.de>",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/visual-framework/vf-core/issues"
+  },
+  "homepage": "https://github.com/visual-framework/vf-core#readme"
+}
